+++ app/assets/current_app_home_block.json
