--- assets/README.md
@@ -1,6 +1,0 @@
-# 各種の設定ファイルや静的コンテンツ
-
-## ファイルの説明
-
-- `manifest.yml`: Slack Appの[manifestファイル](https://api.slack.com/reference/manifests)
-- `current_app_home_block.json`: Slack Appの[App Home](https://api.slack.com/surfaces/app-home)に表示するビュー
