--- assets/manifest.yml
@@ -1,57 +1,0 @@
-display_information:
-  name: RA timecard recorder
-  description: RAの勤怠を記録し、簡単に確認可能にします
-  background_color: "#001657"
-features:
-  app_home:
-    home_tab_enabled: true
-    messages_tab_enabled: false
-    messages_tab_read_only_enabled: true
-  bot_user:
-    display_name: RA timecard recorder
-    always_online: true
-  slash_commands:
-    - command: /init
-      description: ユーザとして初期登録を行う
-      usage_hint: <氏名>
-      should_escape: false
-    - command: /register_ra
-      description: 新しくRAを登録する
-      usage_hint: <RA区分名>
-      should_escape: false
-    - command: /get_working_hours
-      description: 今月の稼働時間を計算する
-      usage_hint: "[yyyy/mm]"
-      should_escape: false
-    - command: /download_csv
-      description: 作業記録をCSV形式でダウンロード
-      usage_hint: "[yyyy/mm]"
-      should_escape: false
-oauth_config:
-  scopes:
-    bot:
-      - app_mentions:read
-      - channels:history
-      - chat:write
-      - commands
-      - files:write
-      - groups:history
-      - groups:read
-      - groups:write
-      - im:history
-      - im:write
-      - mpim:history
-      - im:read
-settings:
-  event_subscriptions:
-    bot_events:
-      - app_mention
-      - message.channels
-      - message.groups
-      - message.im
-      - message.mpim
-  interactivity:
-    is_enabled: true
-  org_deploy_enabled: false
-  socket_mode_enabled: true
-  token_rotation_enabled: false
