--- config.py
@@ -1,9 +1,0 @@
-from typing import Final
-
-APP_HOME_VIEW_FILEPATH: Final[str] = "assets/current_app_home_block.json"
-ADMINISTRATOR_IDs = [
-    "U708JUVPE",  # admin1
-    "U018BCSP5CL",  # admin2
-    "U04947Z9YCE",  # developer on prod server
-    "U058CFA7F7U",  # developer on test server
-]
